pyproject.toml
[tool.poetry]
name = "technicaltestagents"
version = "0.1.0"
description = ""
authors = ["Megha <68321717+MeghaRam-ai@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
swarm = {git = "https://github.com/openai/swarm.git"} #agents
nltk = "^3.9.1"
pandas = "^2.2.3"
fuzzywuzzy = "^0.18.0" #token matching
sentence-transformers = "^3.3.1" #semantic similiarity
python-dotenv = "^1.0.1"
python-levenshtein = "^0.26.1" #for fuzzywuzzy

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
